backend/requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14,<3.15
psycopg2-binary>=2.9,<3.0
django-extensions>=3.2,<3.3
django-tenants>=3.5,<3.6
djangorestframework-simplejwt>=5.3,<5.4
pytest>=7.0,<8.0
pytest-django>=4.5,<5.0
pytest-cov>=4.0,<5.0
